Create own file for dialog fragment tests
